// plantas/src/app/models/planta.ts
export interface Planta {
  id_planta?: number;
  nome_planta?: string;
  id_caracteristica_1_1_1?: number;
  id_ale1_1?: number;
  id_ale2_1?: number;
  id_gene1_1?: number;
  id_gene2_1?: number;
  id_caracteristica_1_1_2?: number;
  id_ale1_2?: number;
  id_ale2_2?: number;
  id_gene1_2?: number;
  id_gene2_2?: number;
  id_caracteristica_1_1_3?: number;
  id_ale1_3?: number;
  id_ale2_3?: number;
  id_gene1_3?: number;
  id_gene2_3?: number;
  id_caracteristica_1_1_4?: number;
  id_ale1_4?: number;
  id_ale2_4?: number;
  id_gene1_4?: number;
  id_gene2_4?: number;
  id_caracteristica_1_1_5?: number;
  id_ale1_5?: number;
  id_ale2_5?: number;
  id_gene1_5?: number;
  id_gene2_5?: number;
  id_caracteristica_1_1_6?: number;
  id_ale1_6?: number;
  id_ale2_6?: number;
  id_gene1_6?: number;
  id_gene2_6?: number;
  id_caracteristica_1_1_7?: number;
  id_ale1_7?: number;
  id_ale2_7?: number;
  id_gene1_7?: number;
  id_gene2_7?: number;
  created_at?: Date;
}